requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
fusionauth-client==1.31.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pkce==1.0.3
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.26
SQLAlchemy-Utils==0.37.9
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
